requirements.txt
# Core dependencies
gradio>=4.0.0
mcp>=1.0.0
python-dotenv>=1.0.0

# Model and AI dependencies
transformers>=4.30.0
torch>=2.0.0
accelerate>=0.20.0
openai>=1.0.0

# For smolagents-based client
smolagents>=0.1.0

# Common utilities
numpy>=1.20.0
json5>=0.9.0
requests>=2.28.0

# Testing dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0

# Development dependencies
pre-commit>=4.0.0
